Check relocation symbols in both .rela.plt and .rela.dyn.
When an executable is compiled with partial RELRO and no-plt, __libc_start_main
is in .rela.plt and all other relocation symbols are in .rela.dyn.
However plthook doesn't check .rela.dyn when .rela.plt is found.
So both .rela.plt and .rela.dyn must be checked to find a relocation entry.